requirements.txt
# 核心依赖
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0
scikit-learn>=1.0.0

# 因果推断
causal-learn>=0.1.3
dowhy>=0.11
CausalML>=0.13.0

# 网络分析
networkx>=2.6.0
graphviz>=0.20.0

# 可视化
matplotlib>=3.5.0
seaborn>=0.12.0
plotly>=5.10.0

# 数据处理
pyyaml>=6.0
openpyxl>=3.0.0  # Excel支持

# Jupyter
jupyter>=1.0.0
ipykernel>=6.0.0
notebook>=6.4.0

# 进度条
tqdm>=4.65.0

# 统计
statsmodels>=0.14.0
pingouin>=0.5.0  # 更高级的统计检验

# LLM（可选，用于文本提取）
openai>=1.0.0
anthropic>=0.18.0
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-core>=0.1.0
pydantic>=2.0.0

# 其他
python-dotenv>=1.0.0
